requirements.txt
django==2.0
django-widget-tweaks
django-adminlte3
django-filter
mysqlclient
